Removed commented code and fixed comments
