+++ engine/common/rpc/convert/collections_test.go
@@ -94,58 +94,3 @@ func TestConvertCollectionGuarantees(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, guarantees, converted)
 }
-
-func TestNewCollectionGuarantee(t *testing.T) {
-	t.Run("valid guarantee", func(t *testing.T) {
-		ug := flow.UntrustedCollectionGuarantee{
-			CollectionID:     unittest.IdentifierFixture(),
-			ReferenceBlockID: unittest.IdentifierFixture(),
-			ChainID:          flow.Testnet,
-			SignerIndices:    []byte{0, 1, 2},
-			Signature:        unittest.SignatureFixture(),
-		}
-
-		guarantee, err := flow.NewCollectionGuarantee(ug)
-		assert.NoError(t, err)
-		assert.NotNil(t, guarantee)
-	})
-
-	t.Run("missing collection ID", func(t *testing.T) {
-		ug := flow.UntrustedCollectionGuarantee{
-			CollectionID:     flow.ZeroID,
-			ReferenceBlockID: unittest.IdentifierFixture(),
-			SignerIndices:    []byte{1},
-		}
-
-		guarantee, err := flow.NewCollectionGuarantee(ug)
-		assert.Error(t, err)
-		assert.Nil(t, guarantee)
-		assert.Contains(t, err.Error(), "CollectionID")
-	})
-
-	t.Run("missing reference block ID", func(t *testing.T) {
-		ug := flow.UntrustedCollectionGuarantee{
-			CollectionID:     unittest.IdentifierFixture(),
-			ReferenceBlockID: flow.ZeroID,
-			SignerIndices:    []byte{1},
-		}
-
-		guarantee, err := flow.NewCollectionGuarantee(ug)
-		assert.Error(t, err)
-		assert.Nil(t, guarantee)
-		assert.Contains(t, err.Error(), "ReferenceBlockID")
-	})
-
-	t.Run("missing signer indices", func(t *testing.T) {
-		ug := flow.UntrustedCollectionGuarantee{
-			CollectionID:     unittest.IdentifierFixture(),
-			ReferenceBlockID: unittest.IdentifierFixture(),
-			SignerIndices:    []byte{},
-		}
-
-		guarantee, err := flow.NewCollectionGuarantee(ug)
-		assert.Error(t, err)
-		assert.Nil(t, guarantee)
-		assert.Contains(t, err.Error(), "SignerIndices")
-	})
-}
