+++ engine/execution/ingestion/uploader/retryable_uploader_wrapper.go
@@ -217,13 +217,6 @@ func (b *BadgerRetryableUploaderWrapper) reconstructComputationResult(
 			Guarantee:  guarantees[inx],
 			Collection: collection,
 		}
-
-		//completeCollections[collectionID] = &entity.CompleteCollection{
-		//	Guarantee: guarantees[inx],
-		//	Collection: &flow.Collection{
-		//		Transactions: collection.Transactions,
-		//	},
-		//}
 	}
 
 	// retrieving TransactionResults from BadgerDB
