+++ model/flow/collectionGuarantee.go
@@ -15,7 +15,7 @@ type CollectionGuarantee struct {
 	ReferenceBlockID Identifier       // defines expiry of the collection
 	ChainID          ChainID          // the chainID of the cluster in order to determine which cluster this guarantee belongs to
 	SignerIndices    []byte           // encoded indices of the signers
-	Signature        crypto.Signature // guarantor signatures
+	Signature        crypto.Signature // guarantor signaturesз
 }
 
 // UntrustedCollectionGuarantee is an untrusted input-only representation of an CollectionGuarantee,
